move commands to new file
