+++ dockerExtension.ts
@@ -17,6 +17,7 @@ import { composeDown, composeRestart, composeUp } from './commands/docker-compos
 import inspectImage from './commands/inspect-image';
 import { openShellContainer } from './commands/open-shell-container';
 import { pushImage } from './commands/push-image';
+import { consolidateDefaultRegistrySettings, setRegistryAsDefault } from './commands/registrySettings';
 import { removeContainer } from './commands/remove-container';
 import { removeImage } from './commands/remove-image';
 import { restartContainer } from './commands/restart-container';
@@ -41,7 +42,6 @@ import { WebAppCreator } from './explorer/deploy/webAppCreator';
 import { DockerExplorerProvider } from './explorer/dockerExplorer';
 import { AzureImageTagNode, AzureRegistryNode, AzureRepositoryNode } from './explorer/models/azureRegistryNodes';
 import { connectCustomRegistry, disconnectCustomRegistry } from './explorer/models/customRegistries';
-import { CustomRegistryNode } from './explorer/models/customRegistryNodes';
 import { DockerHubImageTagNode, DockerHubOrgNode, DockerHubRepositoryNode } from './explorer/models/dockerHubNodes';
 import { browseAzurePortal } from './explorer/utils/azureUtils';
 import { browseDockerHub, dockerHubLogout } from './explorer/utils/dockerHubUtils';
@@ -192,37 +192,6 @@ function registerDockerCommands(azureAccount: AzureAccount): void {
     registerAzureCommand('vscode-docker.delete-ACR-Image', deleteAzureImage);
     registerAzureCommand('vscode-docker.delete-ACR-Repository', deleteRepository);
     registerAzureCommand('vscode-docker.create-ACR-Registry', createRegistry);
-}
-
-async function setRegistryAsDefault(node: CustomRegistryNode | AzureRegistryNode | DockerHubOrgNode): Promise<void> {
-    let registryName: string;
-    if (node instanceof DockerHubOrgNode) {
-        registryName = node.namespace;
-    } else if (node instanceof AzureRegistryNode) {
-        registryName = node.registry.loginServer || node.label;
-    } else { //CustomREgistryNode
-        registryName = node.registryName;
-    }
-    const configOptions: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration('docker');
-    await configOptions.update('defaultRegistryPath', registryName, vscode.ConfigurationTarget.Workspace);
-    vscode.window.showInformationMessage(`Update defaultRegistryPath to ${registryName}`);
-}
-
-async function consolidateDefaultRegistrySettings(): Promise<void> {
-    const configOptions: vscode.WorkspaceConfiguration = vscode.workspace.getConfiguration('docker');
-    let defaultRegistryPath = configOptions.get('defaultRegistryPath', '');
-    let defaultRegistry = configOptions.get('defaultRegistry', '');
-
-    if (defaultRegistry) {
-        if (defaultRegistryPath) { // combine both the settings, then notify the user
-            await configOptions.update('defaultRegistryPath', `${defaultRegistry}/${defaultRegistryPath}`, vscode.ConfigurationTarget.Workspace);
-        }
-        if (!defaultRegistryPath) {// assign defaultRegistry to defaultRegistryPath
-            await configOptions.update('defaultRegistryPath', `${defaultRegistry}`, vscode.ConfigurationTarget.Workspace);
-        }
-        await configOptions.update('defaultRegistry', undefined, vscode.ConfigurationTarget.Workspace);
-        vscode.window.showInformationMessage("The 'docker.defaultRegistry' setting is now obsolete, and you should just use the 'docker.defaultRegistryPath' setting. Your settings have been updated to reflect this change.")
-    }
 }
 
 export async function deactivate(): Promise<void> {
